Form Validation & Reloading
- Built method for reloading appv2 registerMenus into HM Class, cleaned up code in settings menus because of this.
- Move MandatoryFields logic processing to a utility file to match layout of other settings.
